+++ apps/restconf/restconf_lib.c
@@ -685,11 +685,8 @@ restconf_insert_attributes(cxobj *xdata,
 	if (xmlns_set(xdata, "yang", YANG_XML_NAMESPACE) < 0)
 	    goto done;
 	/* Then add insert attribute */
-	if ((xa = xml_new("insert", xdata, NULL)) == NULL)
+	if ((xa = xml_new("insert", "yang", xdata, CX_ATTR)) == NULL)
 	    goto done;
-	if (xml_prefix_set(xa, "yang") < 0)
-	    goto done;
-	xml_type_set(xa, CX_ATTR);
 	if (xml_value_set(xa, instr) < 0)
 	    goto done;
     }
@@ -703,11 +700,8 @@ restconf_insert_attributes(cxobj *xdata,
 	else
 	    attrname="value";
 	/* Then add value/key attribute */
-	if ((xa = xml_new(attrname, xdata, NULL)) == NULL)
+	if ((xa = xml_new(attrname, "yang", xdata, CX_ATTR)) == NULL)
 	    goto done;
-	if (xml_prefix_set(xa, "yang") < 0)
-	    goto done;
-	xml_type_set(xa, CX_ATTR);
 	if ((ret = api_path2xpath(pstr, ys_spec(y), &xpath, &nsc, NULL)) < 0)
 	    goto done;
 	if ((cb = cbuf_new()) == NULL){